Fix CI tests (w/ x-fetch bump)
