+++ package.json
@@ -97,9 +97,9 @@
     "@polkadot/api-base": "^8.12.1",
     "@polkadot/api-contract": "^8.12.1",
     "@polkadot/api-derive": "^8.12.1",
-    "@polkadot/hw-ledger": "^10.0.1",
-    "@polkadot/keyring": "^10.0.1",
-    "@polkadot/networks": "^10.0.1",
+    "@polkadot/hw-ledger": "^10.0.2",
+    "@polkadot/keyring": "^10.0.2",
+    "@polkadot/networks": "^10.0.2",
     "@polkadot/phishing": "^0.16.2",
     "@polkadot/rpc-augment": "^8.12.1",
     "@polkadot/rpc-core": "^8.12.1",
@@ -110,8 +110,8 @@
     "@polkadot/types-create": "^8.12.1",
     "@polkadot/types-known": "^8.12.1",
     "@polkadot/types-support": "^8.12.1",
-    "@polkadot/util": "^10.0.1",
-    "@polkadot/util-crypto": "^10.0.1",
+    "@polkadot/util": "^10.0.2",
+    "@polkadot/util-crypto": "^10.0.2",
     "@polkadot/wasm-crypto": "^6.2.3",
     "babel-core": "^7.0.0-bridge.0",
     "styled-components": "^5.3.1",
