+++ packages/apps-config/package.json
@@ -33,10 +33,10 @@
     "@phala/typedefs": "0.2.30",
     "@polkadot/api": "^8.12.1",
     "@polkadot/api-derive": "^8.12.1",
-    "@polkadot/networks": "^10.0.1",
+    "@polkadot/networks": "^10.0.2",
     "@polkadot/types": "^8.12.1",
-    "@polkadot/util": "^10.0.1",
-    "@polkadot/x-fetch": "^10.0.1",
+    "@polkadot/util": "^10.0.2",
+    "@polkadot/x-fetch": "^10.0.2",
     "@polymathnetwork/polymesh-types": "0.0.2",
     "@snowfork/snowbridge-types": "0.2.7",
     "@sora-substrate/type-definitions": "1.9.13",
