+++ packages/apps-config/src/ci/runner.ts
@@ -1,18 +1,101 @@
 // Copyright 2017-2022 @polkadot/apps-config authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { isString } from '@polkadot/util';
+import fs from 'fs';
 
+import { ApiPromise, WsProvider } from '@polkadot/api';
+import { assert, isError, isString } from '@polkadot/util';
+
+import { typesBundle, typesChain } from '../api';
 import { createWsEndpoints } from '../endpoints';
-import { checkEndpoint } from './check';
+import { fetchJson } from './fetch';
 
 interface Endpoint {
   name: string;
   ws: string;
 }
 
+interface DnsResponse {
+  Answer?: { name: string }[];
+  Question: { name: string }[];
+}
+
+const TICK = '`';
+
+function createChecker (issueFile: string, failures: string[]): (name: string, ws: string) => Promise<boolean> {
+  return async (name: string, ws: string): Promise<boolean> => {
+    const [,, hostWithPort] = ws.split('/');
+    const [host] = hostWithPort.split(':');
+    const json = await fetchJson<DnsResponse>(`https://dns.google/resolve?name=${host}`);
+
+    let provider: WsProvider | null = null;
+    let api: ApiPromise | null = null;
+    let timerId: NodeJS.Timeout | null = null;
+    let isOk = true;
+
+    try {
+      assert(json.Answer, `No DNS entry for ${host}`);
+
+      provider = new WsProvider(ws, false);
+      api = new ApiPromise({
+        provider,
+        throwOnConnect: true,
+        throwOnUnknown: false,
+        typesBundle,
+        typesChain
+      });
+
+      setTimeout((): void => {
+        provider &&
+          provider
+            .connect()
+            .catch(() => undefined);
+      }, 1000);
+
+      await Promise.race([
+        // eslint-disable-next-line promise/param-names
+        new Promise((_, reject): void => {
+          timerId = setTimeout((): void => {
+            timerId = null;
+            reject(new Error(`Timeout connecting to ${ws}`));
+          }, 30_000);
+        }),
+        api.isReadyOrError
+          .then((a) => a.rpc.chain.getBlock())
+          .then((b) => console.log(`>>> ${name} @ ${ws}`, b.toHuman()))
+      ]);
+    } catch (error) {
+      if (isError(error) && failures.some((f) => (error as Error).message.includes(f))) {
+        process.env.CI_LOG && fs.appendFileSync(issueFile, `\n${TICK}${name} @ ${ws} ${error.message}${TICK}\n`);
+
+        isOk = false;
+      }
+
+      console.error(JSON.stringify(error));
+    } finally {
+      if (timerId) {
+        clearTimeout(timerId);
+      }
+
+      if (provider) {
+        try {
+          if (api) {
+            await api.disconnect();
+          } else {
+            await provider.disconnect();
+          }
+        } catch {
+          // ignore
+        }
+      }
+    }
+
+    return isOk;
+  };
+}
+
 export function checkEndpoints (issueFile: string, failures: string[]): void {
-  const checker = checkEndpoint(issueFile, failures);
+  const checker = createChecker(issueFile, failures);
 
   it.each(
     createWsEndpoints()
@@ -29,9 +112,7 @@ export function checkEndpoints (issueFile: string, failures: string[]): void {
         ws: value
       }))
       .filter((v): v is Endpoint => !!v.ws)
-  )('%name @ %$ws', ({ name, ws }): Promise<void> => {
-    console.error(`>>> ${name} @ ${ws}`);
-
-    return checker(name, ws).then((r) => expect(r).toBe(true));
+  )('%name @ %$ws', async ({ name, ws }): Promise<void> => {
+    expect(await checker(name, ws)).toBe(true);
   });
 }
