+++ packages/page-accounts/package.json
@@ -14,11 +14,11 @@
   "version": "0.116.2-94-x",
   "dependencies": {
     "@babel/runtime": "^7.18.6",
-    "@polkadot/hw-ledger": "^10.0.1",
+    "@polkadot/hw-ledger": "^10.0.2",
     "@polkadot/phishing": "^0.16.2",
     "@polkadot/react-components": "^0.116.2-94-x",
     "@polkadot/react-hooks": "^0.116.2-94-x",
-    "@polkadot/util": "^10.0.1",
+    "@polkadot/util": "^10.0.2",
     "@polkadot/vanitygen": "^0.51.15",
     "detect-browser": "^5.3.0",
     "file-saver": "^2.0.5"
