+++ packages/react-components/package.json
@@ -17,15 +17,15 @@
     "@fortawesome/fontawesome-svg-core": "^6.1.1",
     "@fortawesome/free-solid-svg-icons": "^6.1.1",
     "@fortawesome/react-fontawesome": "^0.1.19",
-    "@polkadot/keyring": "^10.0.1",
+    "@polkadot/keyring": "^10.0.2",
     "@polkadot/react-api": "^0.116.2-94-x",
     "@polkadot/react-identicon": "^2.8.1",
     "@polkadot/react-qr": "^2.8.1",
     "@polkadot/react-query": "^0.116.2-94-x",
     "@polkadot/ui-keyring": "^2.8.1",
     "@polkadot/ui-settings": "^2.8.1",
-    "@polkadot/util": "^10.0.1",
-    "@polkadot/util-crypto": "^10.0.1",
+    "@polkadot/util": "^10.0.2",
+    "@polkadot/util-crypto": "^10.0.2",
     "chart.js": "^3.8.0",
     "codeflask": "^1.4.1",
     "i18next": "^21.8.11",
