+++ packages/react-hooks/package.json
@@ -14,7 +14,7 @@
   "version": "0.116.2-94-x",
   "dependencies": {
     "@babel/runtime": "^7.18.6",
-    "@polkadot/hw-ledger": "^10.0.1",
+    "@polkadot/hw-ledger": "^10.0.2",
     "is-ipfs": "6.0.2"
   }
 }
